Update versionNumber field type from BigInteger to Long in DefendantAccountEntity tests

+++ src/test/java/uk/gov/hmcts/opal/service/opal/OpalImpositionServiceTest.java
@@ -46,7 +46,7 @@ class OpalImpositionServiceTest {
 
         DefendantAccountEntity account = DefendantAccountEntity.builder()
             .defendantAccountId(defendantAccountId)
-              .versionNumber(BigInteger.valueOf(12L))
+              .versionNumber(12L)
             .build();
         when(defendantAccountRepositoryService.findById(defendantAccountId)).thenReturn(account);
         when(impositionRepository.findDefendantAccountImpositionsByDefendantAccountId(defendantAccountId))
